Rename source directory and use snake case for file names
