+++ algaeplot/gui/collapsible_box.py
