+++ algaeplot/gui/configuration.py
