+++ algaeplot/gui/data_list_item.py
