+++ algaeplot/gui/error_window.py
