--- test/files/150.txt
@@ -1,449 +1,0 @@
-[Header]
-Version="5.3"
-Program="Algem"
-Mode="Logged"
-Date="2018/01/16"
-Time="18:51:12"
-Setup="C:\ProgramData\Algenuity\Algem 1.4\Profiles\Lydia\Testing C1\20_150_light_c1.algp"
-Title="400 c1"
-Sensors=1
-Sensors0="Channel=OD,Sensor=Absorbance,BWlower=0,BWupper=1,Name=Optical Density"
-Controls=1
-Controls0="Curve="
-Stats="Serial=13134B,ReactorName=13134B,Curve="
-Axes=2
-Axes0="Name=Time (s),Min=0,Max=86400,Major=10800,Minor=3600,Format=Time,Precision=0,Map=Lin"
-Axes1="Name=OD (Numeric),Min=0,Max=6,Major=1,Minor=0.2,Format=Dec,Precision=0,Map=Lin"
-Markers=0
-Cursors=0
-Horiz="Name=Time,Unit=s,Map=Lin,Start=0,Step=0,Points=0"
-Signals=1
-Signals0="Name=OD,Unit=Numeric,Range=6,Display=1,Position=0,Axis=1"
-[Data]
-1	0.078
-1801	0.075
-3601	0.076
-5401	0.073
-7201	0.075
-9001	0.08
-10801	0.075
-12601	0.078
-14401	0.08
-16201	0.081
-18001	0.085
-19801	0.082
-21601	0.087
-23401	0.079
-25201	0.085
-27001	0.083
-28801	0.089
-30601	0.086
-32401	0.089
-34201	0.083
-36001	0.089
-37801	0.089
-39601	0.09
-41401	0.09
-43201	0.092
-45001	0.09
-46801	0.093
-48601	0.091
-50401	0.089
-52201	0.093
-54001	0.092
-55801	0.088
-57601	0.093
-59401	0.094
-61201	0.092
-63001	0.093
-64801	0.095
-66601	0.097
-68401	0.095
-70201	0.104
-72001	0.104
-73801	0.096
-75601	0.099
-77401	0.102
-79201	0.103
-81001	0.108
-82801	0.111
-84601	0.108
-86401	0.111
-88201	0.116
-90001	0.119
-91801	0.114
-93601	0.118
-95401	0.114
-97201	0.115
-99001	0.118
-100801	0.124
-102601	0.12
-104401	0.124
-106201	0.12
-108001	0.128
-109801	0.128
-111601	0.125
-113401	0.126
-115201	0.119
-117001	0.125
-118801	0.126
-120601	0.129
-122401	0.128
-124201	0.138
-126001	0.132
-127801	0.132
-129601	0.136
-131401	0.132
-133201	0.141
-135001	0.139
-136801	0.137
-138601	0.134
-140401	0.132
-142201	0.136
-144001	0.141
-145801	0.137
-147601	0.151
-149401	0.149
-151201	0.145
-153001	0.151
-154801	0.148
-156601	0.154
-158401	0.147
-160201	0.152
-162001	0.154
-163801	0.156
-165601	0.157
-167401	0.154
-169201	0.161
-171001	0.163
-172801	0.162
-174601	0.159
-176401	0.159
-178201	0.174
-180001	0.175
-181801	0.17
-183601	0.167
-185401	0.168
-187201	0.182
-189001	0.18
-190801	0.184
-192601	0.177
-194401	0.182
-196201	0.182
-198001	0.183
-199801	0.192
-201601	0.197
-203401	0.194
-205201	0.195
-207001	0.193
-208801	0.195
-210601	0.201
-212401	0.194
-214201	0.199
-216001	0.203
-217801	0.21
-219601	0.222
-221401	0.207
-223201	0.225
-225001	0.192
-226801	0.216
-228601	0.225
-230401	0.234
-232201	0.231
-234001	0.218
-235801	0.234
-237601	0.238
-239401	0.233
-241201	0.247
-243001	0.246
-244801	0.256
-246601	0.251
-248401	0.242
-250201	0.257
-252001	0.246
-253801	0.268
-255601	0.271
-257401	0.273
-259201	0.275
-261001	0.269
-262801	0.273
-264601	0.28
-266401	0.276
-268201	0.294
-270001	0.273
-271801	0.29
-273601	0.309
-275401	0.293
-277201	0.306
-279001	0.308
-280801	0.299
-282601	0.316
-284401	0.322
-286201	0.34
-288001	0.318
-289801	0.323
-291601	0.328
-293401	0.342
-295201	0.323
-297001	0.346
-298801	0.341
-300601	0.358
-302401	0.367
-304201	0.362
-306001	0.34
-307801	0.344
-309601	0.347
-311401	0.365
-313201	0.403
-315001	0.383
-316801	0.399
-318601	0.36
-320401	0.408
-322201	0.397
-324001	0.415
-325801	0.393
-327601	0.429
-329401	0.417
-331201	0.42
-333001	0.449
-334801	0.431
-336601	0.443
-338401	0.456
-340201	0.423
-342001	0.464
-343801	0.454
-345601	0.468
-347401	0.47
-349201	0.476
-351001	0.478
-352801	0.486
-354601	0.484
-356401	0.461
-358201	0.449
-360001	0.48
-361801	0.501
-363601	0.513
-365401	0.511
-367201	0.516
-369001	0.547
-370801	0.524
-372601	0.591
-374401	0.589
-376201	0.588
-378001	0.595
-379801	0.577
-381601	0.6
-383401	0.632
-385201	0.62
-387001	0.622
-388801	0.666
-390601	0.606
-392401	0.648
-394201	0.624
-396001	0.645
-397801	0.678
-399601	0.614
-401401	0.654
-403201	0.714
-405001	0.699
-406801	0.691
-408601	0.707
-410401	0.741
-412201	0.718
-414001	0.608
-415801	0.76
-417601	0.729
-419401	0.754
-421201	0.738
-423001	0.622
-424801	0.768
-426601	0.761
-428401	0.831
-430201	0.837
-432001	0.822
-433801	0.851
-435601	0.826
-437401	0.908
-439201	0.876
-441001	0.931
-442801	0.781
-444601	0.943
-446401	0.941
-448201	0.837
-450001	0.899
-451801	0.896
-453601	0.997
-455401	1.032
-457201	0.949
-459001	1.043
-460801	1.04
-462601	0.917
-464401	1.044
-466201	1.002
-468001	1.044
-469801	1.001
-471601	1.081
-473401	1.04
-475201	1.138
-477001	1.034
-478801	1.062
-480601	1.077
-482401	1.12
-484201	1.029
-486001	1.218
-487801	1.135
-489601	1.233
-491401	1.233
-493201	1.196
-495001	1.226
-496801	1.245
-498601	1.079
-500401	1.273
-502201	1.222
-504001	1.321
-505801	1.27
-507601	1.345
-509401	1.366
-511201	1.285
-513001	1.308
-514801	1.344
-516601	1.399
-518401	1.37
-520201	1.37
-522001	1.333
-523801	1.436
-525601	1.305
-527401	1.454
-529201	1.461
-531001	1.421
-532801	1.346
-534601	1.531
-536401	1.461
-538201	1.534
-540001	1.485
-541801	1.476
-543601	1.511
-545401	1.425
-547201	1.579
-549001	1.53
-550801	1.651
-552601	1.508
-554401	1.594
-556201	1.586
-558001	1.598
-559801	1.653
-561601	1.67
-563401	1.525
-565201	1.655
-567001	1.719
-568801	1.648
-570601	1.724
-572401	1.689
-574201	1.768
-576001	1.771
-577801	1.802
-579601	1.797
-581401	1.82
-583201	1.87
-585001	1.822
-586801	1.847
-588601	1.861
-590401	1.906
-592201	1.813
-594001	1.797
-595801	1.935
-597601	1.861
-599401	1.931
-601201	1.901
-603001	1.844
-604801	1.893
-606601	1.976
-608401	1.98
-610201	2.001
-612001	1.979
-613801	2.041
-615601	1.956
-617401	2.035
-619201	1.996
-621001	2.059
-622801	2.033
-624601	2.123
-626401	1.985
-628201	2.123
-630001	2.156
-631801	2.115
-633601	2.118
-635401	2.058
-637201	2.13
-639001	2.2
-640801	2.147
-642601	2.147
-644401	2.174
-646201	2.187
-648001	2.148
-649801	2.126
-651601	2.244
-653401	2.259
-655201	2.232
-657001	2.24
-658801	2.238
-660601	2.293
-662401	2.301
-664201	2.27
-666001	2.317
-667801	2.363
-669601	2.294
-671401	2.349
-673201	2.313
-675001	2.376
-676801	2.379
-678601	2.352
-680401	2.347
-682201	2.391
-684001	2.376
-685801	2.428
-687601	2.387
-689401	2.437
-691201	2.47
-693001	2.422
-694801	2.434
-696601	2.443
-698401	2.461
-700201	2.365
-702001	2.438
-703801	2.53
-705601	2.402
-707401	2.458
-709201	2.547
-711001	2.496
-712801	2.458
-714601	2.518
-716401	2.528
-718201	2.467
-720001	2.506
-721801	2.534
-723601	2.569
-725401	2.604
-727201	2.551
-729001	2.583
-730801	2.556
-732601	2.625
-734401	2.548
-736201	2.63
-738001	2.644
-739801	2.579
-741601	2.627
-743401	2.697
-745201	2.63
-747001	2.657
-748801	2.66
-750601	2.65
-752401	2.665
-754201	2.678
-756001	2.673
-757801	2.615
-759601	2.699
-761401	2.721
-763201	2.704
-765001	2.688
-[End]
