+++ test/files/Algem-HT24/19775 Details.csv
@@ -1,0 +1,53 @@
+Date,2020/02/13
+Time,14:51:25
+Experiment Name,
+Serial Number,19775
+Setup File,Sao ptxD.algt
+A1 Profile,C:\ProgramData\Algenuity\AlgemHT24 1.3\Profiles\Chlamy_constant.algp
+A2 Profile,C:\ProgramData\Algenuity\AlgemHT24 1.3\Profiles\Chlamy_constant.algp
+A3 Profile,C:\ProgramData\Algenuity\AlgemHT24 1.3\Profiles\Chlamy_constant.algp
+A4 Profile,C:\ProgramData\Algenuity\AlgemHT24 1.3\Profiles\Chlamy_constant.algp
+A5 Profile,C:\ProgramData\Algenuity\AlgemHT24 1.3\Profiles\Chlamy_constant.algp
+A6 Profile,C:\ProgramData\Algenuity\AlgemHT24 1.3\Profiles\Chlamy_constant.algp
+B1 Profile,C:\ProgramData\Algenuity\AlgemHT24 1.3\Profiles\Chlamy_constant.algp
+B2 Profile,C:\ProgramData\Algenuity\AlgemHT24 1.3\Profiles\Chlamy_constant.algp
+B3 Profile,C:\ProgramData\Algenuity\AlgemHT24 1.3\Profiles\Chlamy_constant.algp
+B4 Profile,C:\ProgramData\Algenuity\AlgemHT24 1.3\Profiles\Chlamy_constant.algp
+B5 Profile,C:\ProgramData\Algenuity\AlgemHT24 1.3\Profiles\Chlamy_constant.algp
+B6 Profile,C:\ProgramData\Algenuity\AlgemHT24 1.3\Profiles\Chlamy_constant.algp
+C1 Profile,C:\ProgramData\Algenuity\AlgemHT24 1.3\Profiles\Chlamy_constant.algp
+C2 Profile,C:\ProgramData\Algenuity\AlgemHT24 1.3\Profiles\Chlamy_constant.algp
+C3 Profile,C:\ProgramData\Algenuity\AlgemHT24 1.3\Profiles\Chlamy_constant.algp
+C4 Profile,C:\ProgramData\Algenuity\AlgemHT24 1.3\Profiles\Chlamy_constant.algp
+C5 Profile,C:\ProgramData\Algenuity\AlgemHT24 1.3\Profiles\Chlamy_constant.algp
+C6 Profile,C:\ProgramData\Algenuity\AlgemHT24 1.3\Profiles\Chlamy_constant.algp
+D1 Profile,C:\ProgramData\Algenuity\AlgemHT24 1.3\Profiles\Chlamy_constant.algp
+D2 Profile,C:\ProgramData\Algenuity\AlgemHT24 1.3\Profiles\Chlamy_constant.algp
+D3 Profile,C:\ProgramData\Algenuity\AlgemHT24 1.3\Profiles\Chlamy_constant.algp
+D4 Profile,C:\ProgramData\Algenuity\AlgemHT24 1.3\Profiles\Chlamy_constant.algp
+D5 Profile,C:\ProgramData\Algenuity\AlgemHT24 1.3\Profiles\Chlamy_constant.algp
+D6 Profile,C:\ProgramData\Algenuity\AlgemHT24 1.3\Profiles\Chlamy_constant.algp
+A1 Replicates, C3 C4
+A2 Replicates, D3 D5
+A3 Replicates, B2 C1
+A4 Replicates, B6 D1
+A5 Replicates, B3 D2
+A6 Replicates, B1 C6
+B1 Replicates, A6 C6
+B2 Replicates, A3 C1
+B3 Replicates, A5 D2
+B4 Replicates, C5 D6
+B5 Replicates, C2 D4
+B6 Replicates, A4 D1
+C1 Replicates, A3 B2
+C2 Replicates, B5 D4
+C3 Replicates, A1 C4
+C4 Replicates, A1 C3
+C5 Replicates, B4 D6
+C6 Replicates, A6 B1
+D1 Replicates, A4 B6
+D2 Replicates, A5 B3
+D3 Replicates, A2 D5
+D4 Replicates, B5 C2
+D5 Replicates, A2 D3
+D6 Replicates, B4 C5
